End of Project

+++ BlogSystemDotNetClient/BlogSystemDotNetClient/Controllers/BlogController.cs
@@ -10,13 +10,31 @@ namespace BlogSystemDotNetClient.Controllers
     public class BlogController : Controller
     {
         // GET: Blog
+
+        public bool Check()
+        {
+            if(Session["uid"]==null)
+            {
+                return true;
+            }
+            return false;
+        }
+
         public ActionResult AddBlog()
         {
+            if (Check())
+            {
+                return RedirectToAction("Index","Login");
+            }
             return View();
         }
 
         public ActionResult Blog(string title , string content , string blogby)
         {
+            if (Check())
+            {
+                return RedirectToAction("Index", "Login");
+            }
             BlogClient service = new BlogClient();            
             MyBlog stu = new MyBlog();
             stu.blogby = blogby;
@@ -28,6 +46,10 @@ namespace BlogSystemDotNetClient.Controllers
         }
         public ActionResult AllBlog()
         {
+            if (Check())
+            {
+                return RedirectToAction("Index", "Login");
+            }
             BlogClient service = new BlogClient();
             MyBlog[] arr = service.GetAllBlog();
             return View(arr);
@@ -35,6 +57,10 @@ namespace BlogSystemDotNetClient.Controllers
 
         public ActionResult AllMyBlog()
         {
+            if (Check())
+            {
+                return RedirectToAction("Index", "Login");
+            }
             BlogClient service = new BlogClient();
             String id = (String)Session["unm"];
             MyBlog[] arr = service.GetAllMyBlog(id);
@@ -43,6 +69,10 @@ namespace BlogSystemDotNetClient.Controllers
 
         public ActionResult View(string para)
         {
+            if (Check())
+            {
+                return RedirectToAction("Index", "Login");
+            }
             Int32 id = Int32.Parse(para);
             BlogClient service = new BlogClient();
             MyBlog stu = service.GetBlogById(id);
@@ -51,6 +81,10 @@ namespace BlogSystemDotNetClient.Controllers
 
         public ActionResult ViewMy(string para)
         {
+            if (Check())
+            {
+                return RedirectToAction("Index", "Login");
+            }
             Int32 id = Int32.Parse(para);
             BlogClient service = new BlogClient();
             MyBlog stu = service.GetBlogById(id);
@@ -59,6 +93,10 @@ namespace BlogSystemDotNetClient.Controllers
 
         public ActionResult Update(string updateblogNO )
         {
+            if (Check())
+            {
+                return RedirectToAction("Index", "Login");
+            }
             Int32 id = Int32.Parse(updateblogNO);
             BlogClient service = new BlogClient();
             MyBlog stu = service.GetBlogById(id);
@@ -67,6 +105,10 @@ namespace BlogSystemDotNetClient.Controllers
 
         public ActionResult Delete(string deleteblogNO)
         {
+            if (Check())
+            {
+                return RedirectToAction("Index", "Login");
+            }
             Int32 id = Int32.Parse(deleteblogNO);
             BlogClient service = new BlogClient();
             string message = service.DeleteBlog(id);
@@ -75,7 +117,11 @@ namespace BlogSystemDotNetClient.Controllers
 
         public ActionResult UpdateBlog(string blog , string title , string content )
         {
-            
+            if (Check())
+            {
+                return RedirectToAction("Index", "Login");
+            }
+
             BlogClient service = new BlogClient();
             MyBlog stu = new MyBlog();
             stu.blogNO = Int32.Parse(blog);
